project/requirements.txt
Django==4.1.4
environs~=9.5.0
djangorestframework==3.14.0
celery==5.2.7
redis==3.5.3


